+++ packages/ui/src/demo/experiment-view.tsx
@@ -5,7 +5,7 @@ import {
   selectDataPoints,
   useExperiment,
   useSelector,
-} from '@process-optimizer-frontend/core'
+} from '@boostv/process-optimizer-frontend-core'
 
 import catapult from '@ui/testing/sample-data/catapult.json'
 import { DataPoints, Plots } from '..'