+++ packages/ui/src/features/input-model/categorical-variable.tsx
@@ -5,7 +5,7 @@ import { useForm } from 'react-hook-form'
 import CategoricalVariableOptions from './categorical-variable-options'
 import { useStyles } from './categorical-variable.style'
 import { validation } from '@ui/common/forms'
-import { CategoricalVariableType } from '@process-optimizer-frontend/core'
+import { CategoricalVariableType } from '@boostv/process-optimizer-frontend-core'
 
 type CategoricalVariableProps = {
   isDisabled: boolean