+++ sample-app/src/components/experiment/dataEntryTab.tsx
@@ -3,11 +3,11 @@ import {
   selectDataPoints,
   useExperiment,
   useSelector,
-} from '@process-optimizer-frontend/core'
+} from '@boostv/process-optimizer-frontend-core'
 import { ExperimentationGuide } from '@sample/components/result-data/experimentation-guide'
-import { DataEntry } from '@process-optimizer-frontend/core'
+import { DataEntry } from '@boostv/process-optimizer-frontend-core'
 import { useGlobal } from '@sample/context/global'
-import { DataPoints } from '@process-optimizer-frontend/ui'
+import { DataPoints } from '@boostv/process-optimizer-frontend-ui'
 
 export const DataEntryTab = () => {
   const {